PlayerControls  out of spline
Scene: DemoMapJoostScene

Controls:
Press space to start
Hold R to "blow"
Press F to Fly again after landing. To take off again, Hold R and then press F.
When mid-air, after flying for the second time (in the same session) , press G to reset the values to be able to collide with the ground again.